requirements.txt
aos==0.1.2
asgiref==3.4.1
beautifulsoup4==4.9.3
bootstrap4==0.1.0
boto3==1.18.1
botocore==1.21.1
certifi==2021.5.30
dj-database-url==0.5.0
Django==3.2.5
django-bootstrap-form==3.4
django-bootstrap-v5==1.0.0
django-ckeditor==6.1.0
django-crispy-forms==1.11.1
django-js-asset==1.2.2
django-less==0.7.2
django-livereload-server==0.3.2
django-sass==1.0.1
django-sass-processor==1.0.1
django-static-precompiler==2.0
django-storages==1.11.1
gunicorn==20.1.0
install==1.3.4
jedi==0.18.0
jmespath==0.10.0
libsass==0.21.0
parso==0.8.1
Pillow==8.1.2
psycopg2==2.8.6
python-dateutil==2.8.2
pytz==2021.1
rjsmin==1.1.0
s3transfer==0.5.0
six==1.16.0
soupsieve==2.2
sqlparse==0.4.1
tornado==6.1
urllib3==1.26.6
whitenoise==5.3.0
wincertstore==0.2
